Add slug-based routes and update links
