+++ src/components/app/profile/AudioSlider.tsx
@@ -20,7 +20,7 @@ export default function AudioSlider({ posts }: AudioSliderProps) {
         {posts.map((post) => (
           <Link
             key={post.id}
-            href={post.collective_slug ? `/collectives/${post.collective_slug}/${post.id}` : `/posts/${post.id}`}
+            href={post.slug ? `/posts/${post.slug}` : (post.collective_slug ? `/collectives/${post.collective_slug}/${post.id}` : `/posts/${post.id}`)}
             className="shrink-0 w-48 bg-card rounded-xl shadow p-3 hover:ring-2 hover:ring-primary transition"
           >
             <div className="h-24 bg-muted rounded mb-2" />
